requirements.txt
# === Core Framework ===
langchain>=0.2.0
langgraph>=0.1.0
google-generativeai>=0.8.0
google-genai>=1.0.0

# === APIs & Data ===
semanticscholar>=0.8.0
arxiv>=2.1.0
requests>=2.31.0
aiohttp>=3.9.0

# === NLP & Embeddings ===
sentence-transformers>=2.6.0

# === Knowledge Graph ===
networkx>=3.2
python-louvain>=0.16
pyvis>=0.3.2

# === Vector Store ===
chromadb>=0.4.0

# === Analysis ===
scikit-learn>=1.4.0
hdbscan>=0.8.0
umap-learn>=0.5.0
pandas>=2.2.0
numpy>=1.26.0

# === Visualization ===
plotly>=5.20.0
matplotlib>=3.8.0

# === Frontend ===
streamlit>=1.32.0

# === Report Generation ===
fpdf2>=2.7.0
jinja2>=3.1.0

# === PDF Processing ===
pypdf>=4.0.0

# === Utilities ===
python-dotenv>=1.0.0
pydantic>=2.6.0
tenacity>=8.2.0
rich>=13.7.0
